profiles_service/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
Pillow==9.3
#mysql-community-client-plugins
mysql-connector-python==8.0.33
django-mysql==3.9
mysqlclient==2.1.1
django-environ==0.7.0
django-filter==21.1
django-autoslug==1.9.8
django-countries==7.2.1
django-phonenumber-field==5.2.0
phonenumbers==8.12.33
flake8==3.9.2
black==21.6b0
isort==5.9.3
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
pytest-django==4.4.0
pytest-factoryboy==2.1.0
Faker==8.9.1
pytest-cov==2.12.1
redis==3.5.3
celery==4.4.7
flower==0.9.7
django-celery-email==3.0.0
argh==0.26.2
PyYAML==5.3.1
watchdog==1.0.2
django-cors-headers